Finished Magic System Mechanics
With new modifications, the game now has:
- Magic Points
- Health Points
- A Player Management Script (for magic and health and other stuff)

There is a microphone input script that now allows players to recharge magic points by screaming into the microphone.
The louder the input, the faster the magic will charge.

As there will be more spells a few more lines of code will have to be added but the system works regardless.